+++ CHANGELOG.md
@@ -1,18 +1,24 @@
 # Changelog
 
+## [0.2.1] - 2020-08-20Z
+
+### Added
+
+- [bin] Added aliases for each command. ([#21](https://github.com/qryxip/cargo-member/pull/21))
+
 ## [0.2.0] - 2020-08-09Z
 
-## Changed
+### Changed
 
 - [lib, bin] Now `cargo member focus` `deactivate`s existing packages. ([#19](https://github.com/qryxip/cargo-member/pull/19))
 
-## Added
+### Added
 
 - [lib, bin] Added `deactivate` command. ([#18](https://github.com/qryxip/cargo-member/pull/18))
 
 ## [0.1.2] - 2020-08-01Z
 
-## Fixed
+### Fixed
 
 - Now each command `flush` the `stderr` correctly. ([#16](https://github.com/qryxip/cargo-member/pull/16))